feat: App Store pre-submission hardening pass
Backend:
- Add /feed route, feedService (social feed, reactions, split sharing)
- Add block/report infrastructure: user_blocks table migration, blockUser/unblockUser/reportUser in friendService, POST/DELETE /social/block endpoints
- Add /privacy and /terms static HTML routes from public/
- Add use_imperial to profiles (migration 028); RLS enabled on all 29 tables (migration 031)
- Strict Zod schemas for feed payloads, friend/report/block requests; add .strict() to reject unknown keys
- Extract qs()/intParam() helpers to lib/query; apply across all routes
- profileService.deleteAccount() with pg transaction; auth route delegates to service
- feedService uses LEFT JOIN profiles + block-filter subquery
- analyticsService LEFT JOIN exercise_definitions + COALESCE
- sessionService ::float casts; exerciseService default limit 500→50
- Remove dead deps: jsonwebtoken, bcryptjs and their @types
- 63 backend tests passing (5 test files)

iOS:
- Remove Sign in with Apple buttons + entitlement risk; add Forgot Password flow
- @MainActor on AuthStore, AuthViewModel, OnboardingViewModel
- WorkoutSyncService: captures serverID from POST /sessions, pushes sets/finish, reconcile() on launch
- ElosSchema: serverID/syncPending on WorkoutSessionRecord + ExerciseSetRecord; syncPending/useImperial on UserProfileRecord
- ProfileModels.swift: ProfileResponse with use_imperial
- Report + Block in FeedPostCard menu and FriendProfileView toolbar menu
- FeedViewModel: reportPost(), blockAuthor(), loadFailed state, retry UI
- SocialViewModel: blockUser()
- NamedUser protocol + Formatters.swift utilities; dedup displayName/initials/avatarHex across response types
- Extract ViewModels: MachineDetailViewModel, WorkoutDetailViewModel, ExerciseDetailViewModel, FriendProfileViewModel, ReadinessCheckInViewModel
- Deployment target 18.0, iPhone-only (TARGETED_DEVICE_FAMILY=1), portrait-only Debug
- Age floor 13; nutrition disclaimers in onboarding + BodyMetrics
- Legal links → elos.onrender.com; CrewView unfriend via Menu
- 15 iOS unit tests passing

+++ apps/elos-api/src/services/sessionService.ts
@@ -25,7 +25,7 @@ export class SessionService {
        VALUES ($1, $2, $3, $4, $5, $6, $7)
        RETURNING id, user_id,
          started_at::text, finished_at::text,
-         session_rpe, notes, template_id::text, total_volume,
+         session_rpe, notes, template_id::text, total_volume::float,
          created_at::text`,
       [userId, started_at, finished_at, session_rpe, notes, template_id, total_volume]
     );
@@ -36,7 +36,7 @@ export class SessionService {
     const result = await this.db.query<WorkoutSession>(
       `SELECT id, user_id,
          started_at::text, finished_at::text,
-         session_rpe, notes, template_id::text, total_volume,
+         session_rpe, notes, template_id::text, total_volume::float,
          created_at::text
        FROM workout_sessions
        WHERE user_id = $1
@@ -51,7 +51,7 @@ export class SessionService {
     const result = await this.db.query<WorkoutSession>(
       `SELECT id, user_id,
          started_at::text, finished_at::text,
-         session_rpe, notes, template_id::text, total_volume,
+         session_rpe, notes, template_id::text, total_volume::float,
          created_at::text
        FROM workout_sessions
        WHERE id = $1 AND user_id = $2`,
@@ -83,7 +83,7 @@ export class SessionService {
        WHERE id = $${idx} AND user_id = $${idx + 1}
        RETURNING id, user_id,
          started_at::text, finished_at::text,
-         session_rpe, notes, template_id::text, total_volume,
+         session_rpe, notes, template_id::text, total_volume::float,
          created_at::text`,
       values
     );
@@ -113,7 +113,7 @@ export class SessionService {
          (session_id, user_id, exercise_name, set_index, weight_kg, reps, rpe, rir, completed_at)
        VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)
        RETURNING id, session_id, user_id, exercise_name, set_index,
-         weight_kg, reps, rpe, rir,
+         weight_kg::float, reps, rpe::float, rir,
          completed_at::text, created_at::text`,
       [sessionId, userId, exercise_name, set_index, weight_kg, reps, rpe, rir, completed_at]
     );
@@ -123,7 +123,7 @@ export class SessionService {
   async getSessionSets(sessionId: string, userId: string): Promise<ExerciseSet[]> {
     const result = await this.db.query<ExerciseSet>(
       `SELECT id, session_id, user_id, exercise_name, set_index,
-         weight_kg, reps, rpe, rir,
+         weight_kg::float, reps, rpe::float, rir,
          completed_at::text, created_at::text
        FROM exercise_sets
        WHERE session_id = $1 AND user_id = $2
